Extract EditorSidePage into separate files.
